--- lib/controllers/visiting_places_controller.dart
@@ -1,3 +1,0 @@
-import 'package:flutter/material.dart';
-
-class VisitingPlaces extends ChangeNotifier {}
